rename : "domain" -> "model/entity"로 변경 & "type"에 있던 enum파일들을 "model"로 이동
