+++ src/main/java/zb/accountMangement/account/model/entity/Transaction.java
@@ -1,9 +1,9 @@
-package zb.accountMangement.account.domain;
+package zb.accountMangement.account.model.entity;
 
 import lombok.*;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-import zb.accountMangement.account.type.TransactionType;
+import zb.accountMangement.account.model.TransactionType;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
