+++ src/main/java/zb/accountMangement/account/repository/TransactionRepository.java
@@ -2,7 +2,7 @@ package zb.accountMangement.account.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import zb.accountMangement.account.domain.Transaction;
+import zb.accountMangement.account.model.entity.Transaction;
 
 import java.util.List;
