+++ src/main/java/zb/accountMangement/member/repository/MemberRepository.java
@@ -3,7 +3,7 @@ package zb.accountMangement.member.repository;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import zb.accountMangement.member.domain.Member;
+import zb.accountMangement.member.model.entity.Member;
 
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
