+++ src/main/java/zb/accountMangement/stock/controller/StockController.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import zb.accountMangement.stock.domain.StockBalance;
-import zb.accountMangement.stock.domain.Trading;
+import zb.accountMangement.stock.model.entity.StockBalance;
+import zb.accountMangement.stock.model.entity.Trading;
 import zb.accountMangement.stock.dto.*;
 import zb.accountMangement.stock.service.StockService;
 import javax.validation.Valid;
