+++ src/main/java/zb/accountMangement/stock/model/TradeType.java
@@ -1,4 +1,4 @@
-package zb.accountMangement.stock.type;
+package zb.accountMangement.stock.model;
 
 public enum TradeType {
   BUY, SELL,TRANSFER
