+++ src/main/java/zb/accountMangement/stock/repository/StockRepository.java
@@ -2,7 +2,7 @@ package zb.accountMangement.stock.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import zb.accountMangement.stock.domain.Stock;
+import zb.accountMangement.stock.model.entity.Stock;
 
 @Repository
 public interface StockRepository extends JpaRepository<Stock, Long> {
