+++ src/main/java/zb/accountMangement/stock/repository/TradingRepository.java
@@ -2,7 +2,7 @@ package zb.accountMangement.stock.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import zb.accountMangement.stock.domain.Trading;
+import zb.accountMangement.stock.model.entity.Trading;
 
 import java.time.LocalDateTime;
 import java.util.List;
