+++ src/main/java/zb/accountMangement/stock/service/StockService.java
@@ -3,22 +3,22 @@ package zb.accountMangement.stock.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import zb.accountMangement.account.domain.Account;
+import zb.accountMangement.account.model.entity.Account;
 import zb.accountMangement.account.dto.TransactionDto;
 import zb.accountMangement.account.service.AccountService;
 import zb.accountMangement.account.service.TransactionService;
 import zb.accountMangement.common.exception.CustomException;
 import zb.accountMangement.common.type.ErrorCode;
-import zb.accountMangement.stock.domain.Stock;
-import zb.accountMangement.stock.domain.StockBalance;
-import zb.accountMangement.stock.domain.Trading;
+import zb.accountMangement.stock.model.entity.Stock;
+import zb.accountMangement.stock.model.entity.StockBalance;
+import zb.accountMangement.stock.model.entity.Trading;
 import zb.accountMangement.stock.dto.DateDto;
 import zb.accountMangement.stock.dto.TradeStockDto;
 import zb.accountMangement.stock.dto.TransferStockDto;
 import zb.accountMangement.stock.repository.StockBalanceRepository;
 import zb.accountMangement.stock.repository.StockRepository;
 import zb.accountMangement.stock.repository.TradingRepository;
-import zb.accountMangement.stock.type.TradeType;
+import zb.accountMangement.stock.model.TradeType;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
